backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
cryptography==44.0.0
defusedxml==0.7.1
django==4.2.16
django-debug-toolbar==4.4.6
django-filter==24.3
django-stubs==5.1.1
django-stubs-ext==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djangorestframework-stubs==3.15.1
djoser==2.3.1
idna==3.10
isort==5.13.2
oauthlib==3.2.2
Pillow==11.0.0
psycopg==3.2.3
psycopg2-binary==2.9.10
pycparser==2.22
pyjwt==2.10.1
python-dotenv==1.0.1
python3-openid==3.2.0
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.2
tomli==2.2.1
types-pyyaml==6.0.12.20240917
types-requests==2.32.0.20241016
typing-extensions==4.12.2
urllib3==2.2.3
